fixed a minor bug on output csv file

+++ basic_io/h5bench_vpicio.c
@@ -822,14 +822,14 @@ int main(int argc, char* argv[]) {
         printf("Total sleep time %ds, total write size = %lu MB\n", total_sleep_time, total_size_mb);
 
         float rwt_s = (float)raw_write_time / (1000*1000);
-        unsigned long raw_rate_mbs = total_data_size / raw_write_time;
-        printf("RR: Raw write time = %.3f sec, RR = %lu MB/sec \n", rwt_s, raw_rate_mbs);
+        float raw_rate_mbs = (float)total_data_size / raw_write_time;
+        printf("RR: Raw write time = %.3f sec, RR = %.3f MB/sec \n", rwt_s, raw_rate_mbs);
 
         unsigned long meta_time_ms = (t3 - t2 - raw_write_time - sleep_time * (NUM_TIMESTEPS - 1) * 1000 * 1000) / 1000;
         printf("Core metadata time = %lu ms\n", meta_time_ms);
 
-        unsigned long or_mbs = total_data_size / (t4 - t1 - (NUM_TIMESTEPS - 1) * 1000 * 1000);
-        printf("OR (observed rate) = %lu MB/sec\n", or_mbs);
+        float or_mbs = (float)total_data_size / (t4 - t1 - (NUM_TIMESTEPS - 1) * 1000 * 1000);
+        printf("OR (observed rate) = %.3f MB/sec\n", or_mbs);
 
         float oct_s = (float)(t4 - t0) / (1000*1000);
         printf("OCT(observed completion time) = %.3f sec\n", oct_s);
@@ -842,9 +842,9 @@ int main(int argc, char* argv[]) {
             fprintf(params.csv_fs, "Total_sleep_time, %d, sec\n", total_sleep_time);
             fprintf(params.csv_fs, "Total_write_size, %lu, MB\n", total_size_mb);
             fprintf(params.csv_fs, "Raw_write_time, %.3f, sec\n", rwt_s);
-            fprintf(params.csv_fs, "Raw_write_rate, %lu, MB/sec\n", raw_rate_mbs);
+            fprintf(params.csv_fs, "Raw_write_rate, %.3f, MB/sec\n", raw_rate_mbs);
             fprintf(params.csv_fs, "Core_metadata_time, %lu, ms\n", meta_time_ms);
-            fprintf(params.csv_fs, "Observed_rate, %lu, MB/sec\n", or_mbs);
+            fprintf(params.csv_fs, "Observed_rate, %.3f, MB/sec\n", or_mbs);
             fprintf(params.csv_fs, "Observed_completion_time, %.3f, sec\n", oct_s);
             fclose(params.csv_fs);
         }
